3.3- Extracting Chat Controller - #30
