+++ Build-AI-Powered-Apps/4-Project-Building-a-Chatbot/app/packages/server/index.ts
@@ -1,8 +1,7 @@
 import dotenv from "dotenv";
 import type { Request, Response } from "express";
 import express from "express";
-import { z } from "zod";
-import { chatService } from "./services/chat.service";
+import { chatController } from "./controllers/chat.controller";
 
 dotenv.config();
 
@@ -20,35 +19,7 @@ app.get("/api/hello", (req: Request, res: Response) => {
   res.json({ message: "Hello World!" });
 });
 
-const chatSchema = z.object({
-  prompt: z
-    .string()
-    .trim()
-    .min(1, "prompt is required")
-    .max(1000, "prompt is too long, maximum 1000 characters"),
-  conversationId: z.uuid("invalid UUID"),
-});
-
-app.post("/api/chat", async (req: Request, res: Response) => {
-  try {
-    const { prompt, conversationId } = req.body;
-
-    const parsedResult = chatSchema.safeParse(req.body);
-
-    if (!parsedResult.success) {
-      return res.status(400).json(parsedResult.error.format());
-    }
-
-    const response = await chatService.sendMessage(prompt, conversationId);
-    res.json({
-      message: response.message,
-    });
-  } catch (error) {
-    return res.status(500).json({
-      error: "Failed to generate a response",
-    });
-  }
-});
+app.post("/api/chat", chatController.sendMessage);
 
 app.listen(port, () => {
   console.log(`Server is running on http://localhost:${port}`);
